Casted it to float32

+++ keras/train-conv-lstm_natural-scenes.py
@@ -70,25 +70,25 @@ def rolling_window(array, window):
 
 def loadData(data_dir):
 	scenes = h5py.File(data_dir, 'r')
-	stim = np.array(scenes['train/stimulus'])
+	stim = np.array(scenes['train/stimulus']).astype('float32')
 	stim = stim.T
 	X = rolling_window(stim, 40)
 	X = np.rollaxis(X, 2)
 	X = np.rollaxis(X, 3, 1)
 	#Truncate rates to appropriate time frame
-	y = np.array(scenes['train/response/firing_rate_10ms'])
+	y = np.array(scenes['train/response/firing_rate_10ms']).astype('float32')
 	for i in xrange(y.shape[0]): #normalize firing rate of each cell to be between 0 and 1
 		if np.max(y[i]) != 0:
 			y[i] /= np.max(y[i])
 	y = y.T
 	y = y[40:,:]
-	stim_2 = np.array(scenes['test/stimulus'])
+	stim_2 = np.array(scenes['test/stimulus']).astype('float32')
 	stim_2 = stim_2.T
 	X_2 = rolling_window(stim_2, 40)
 	X_2 = np.rollaxis(X_2, 2)
 	X_2 = np.rollaxis(X_2, 3, 1)
 	#Truncate rates to appropriate time frame
-	y_2 = np.array(scenes['test/response/firing_rate_10ms'])
+	y_2 = np.array(scenes['test/response/firing_rate_10ms']).astype('float32')
 	for i in xrange(y_2.shape[0]): #normalize firing rate of each cell to be between 0 and 1
 		if np.max(y_2[i]) != 0:
 			y_2[i] /= np.max(y_2[i])
